refactor(plants): update routing paths and fix imports (step 13)
- Update route paths: plant-form→create, plant-list→management,
  plant-overview→gallery, plant/:id→view/:id, plant-edit/:id→edit/:id
- Fix broken environment imports in components
- Fix model import in plant.service.ts
- Fix background image path in plant-layout component
